change loadparams ram_size to uint64_t.

+++ hw/mips/mips_ls1b.c
@@ -59,7 +59,7 @@ static const int ide_irq[2] = { 14, 15 };
 /* i8254 PIT is attached to the IRQ0 at PIC i8259 */
 
 static struct _loaderparams {
-	int ram_size;
+	int64_t ram_size;
 	const char *kernel_filename;
 	const char *kernel_cmdline;
 	const char *initrd_filename;
